engine/src/main/java/org/cucina/engine/config/TokenRepositoryDefinitionParser.java
package org.cucina.engine.config;

import org.cucina.core.config.AbstractDefaultIdBeanDefinitionParser;
import org.cucina.core.spring.SingletonBeanFactory;
import org.cucina.engine.repository.jpa.TokenRepositoryImpl;
import org.springframework.beans.factory.support.BeanDefinitionBuilder;
import org.w3c.dom.Element;


/**
 * JAVADOC for Class Level
 *
 * @author $Author: $
 * @version $Revision: $
  */
public class TokenRepositoryDefinitionParser
    extends AbstractDefaultIdBeanDefinitionParser {
    private static final String TOKEN_REPO = "tokenRepository";

    /**
    * JAVADOC Method Level Comments
    *
    * @param element JAVADOC.
    *
    * @return JAVADOC.
    */
    @Override
    protected Class<?> getBeanClass(Element element) {
        return TokenRepositoryImpl.class;
    }

    /**
     * JAVADOC Method Level Comments
     *
     * @return JAVADOC.
     */
    @Override
    protected String getDefaultId() {
        return TOKEN_REPO;
    }

    /**
     * JAVADOC Method Level Comments
     *
     * @param element JAVADOC.
     * @param builder JAVADOC.
     */
    @Override
    protected void doParse(Element element, BeanDefinitionBuilder builder) {
        builder.addConstructorArgReference(element.getAttribute(
                SingletonBeanFactory.INSTANCE_FACTORY_ID));
    }
}
